+++ pages/api/auth/[...nextauth].ts
@@ -60,5 +60,6 @@ export default NextAuth({
   },
   pages: {
     signIn: '/login',
+    newUser: '/register',
   },
 });
