+++ pages/home/index.tsx
@@ -1,6 +1,26 @@
-import { HomePage } from '@/pages';
+import { Header } from '@/components/header/header';
+import { Box, Container, Typography } from '@mui/material';
+import Head from 'next/head';
 import React from 'react';
 
 export default function Home(): JSX.Element {
-  return <HomePage />;
+  return (
+    <>
+      <Head>
+        <title>Home</title>
+      </Head>
+      <Container maxWidth="xs">
+        <Header />
+        <Box
+          className={
+            'w-full max-w-screen-xl h-screen flex justify-center items-center'
+          }
+        >
+          <Typography variant="h4" sx={{ my: 3 }}>
+            Página Home
+          </Typography>
+        </Box>
+      </Container>
+    </>
+  );
 }
