+++ src/components/card/index.ts
@@ -1,2 +1,1 @@
 export { default as IntroductionCard } from './introduction-card/card-introduction';
-export { default as ActionAreaCard } from './action-card';
