+++ src/pages/index.ts
@@ -1,9 +1,6 @@
-export { default as HomePage } from './home-page';
 export { default as NotFound } from './error-page/404';
 export { default as RegisterPage } from './account-page/register';
 export { default as LoginPage } from './account-page/login';
 export { default as IntroductionPage } from './introduction-page';
 export { default as RegisterForm } from './account-page/register';
 export { default as LoginForm } from './account-page/login';
-export { default as AdminPage } from './admin';
-export { default as RegisterProductPage } from './register-product';
